Improve TypeFactory cache performance
`TypeFactory.constructSpecializedType` adds elements to the
cache which contain `PlaceholderForType` type arguments. These
type arguments all have the same hashCode (based on Object.class
plus a hard-coded value for all placeholders) while overriding
`equals` to perform reference equality. Each time that
`constructSpecializedType` is called, more data is added to the
cache with the same hash code, but equality never succeeds due to
instances differing, not only reducing cache hit probability, but
making it increasingly difficult to find the values we want due
to following pointers on heavy cache collisoins where the data
structure degenerates into a linked list.
I've committed a change which bypasses the cache in cases where
PlaceholderForType is used. I don't know if this is an ideal
solution, but it improves the case wher I've encountered this
edge case.

Secondly, the hashCode implementations in several components
lead to hash collisions due to commutativity (summation and
bitwise-xor). Take MapLikeType, for example: Map<A, B> would
always have the same hashCode as Map<B, A> due to using
`A.hashCode ^ B.hashCode`. Furthermore, all maps using the same
type as a key and value have the same hashCode due to
`A.hashCode ^ A.hashCode` evaluating to zero.

Lastly, TypeBindings equals and hashCode only took into account
`_types`, but not `_names` or `_unboundVariables`. I've updated
equals and hashCode to include these additional pieces of data,
and to include a fast-path checking the memoized _hashCode first
to avoid unnecessary computation when we can quickly rule out
equality.